move restcountries.com http logic to a separate service - more in description
it might be better to use a CountriesHttpInterface and then map the appropriate concrete implementation based on the 3rd party used.
This way, we can switch out implementation without touching CountrySyncService
