+++ src/Service/Contract/CountriesHttpInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Service\Contract;
+
+interface CountriesHttpInterface
+{
+    public function getRestCountries(): array;
+}
